requirements.txt
discord
Flask
flask-SQLAlchemy
flask-login
colorama